requirements.txt
aioredis==1.3.1
amqp==5.0.6
asgi-redis
asgiref
async-timeout==3.0.1
attrs==21.2.0
autobahn==21.3.1
Automat==20.2.0
billiard==3.6.4.0
celery==5.1.2
certifi==2020.12.5
cffi==1.15.0
channels==3.0.4
channels-redis==3.3.1
chardet==4.0.0
click==7.1.2
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
constantly==15.1.0
cryptography==35.0.0
daphne==3.0.2
dj-database-url==0.5.0
Django==3.2.4
django-cors-headers==3.7.0
django-heroku==0.3.1
django-redis==5.0.0
djangorestframework==3.12.4
gunicorn
hiredis==2.0.0
hyperlink==21.0.0
idna==2.10
incremental==21.3.0
kombu==5.2.0
msgpack==1.0.2
msgpack-python==0.5.6
Pillow==8.2.0
prompt-toolkit==3.0.22
psycopg2==2.8.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pyOpenSSL==21.0.0
python-decouple==3.4
pytz==2021.1
redis==3.5.3
requests==2.25.1
service-identity==21.1.0
six==1.16.0
sqlparse==0.4.1
Twisted==21.7.0
txaio==21.2.1
typing-extensions==3.10.0.2
urllib3==1.26.5
vine==5.0.0
wcwidth==0.2.5
whitenoise==5.2.0
PyJWT
apns2
django-push-notifications==2.0.0
h2==2.6.2
hpack==3.0.0
http-ece==1.1.0
hyper==0.7.0
hyperframe==3.2.0
py-vapid==1.8.2
pywebpush==1.14.0

MarkupSafe
coreapi
coreschema
drf-yasg
inflection
itypes
jinja2
packaging
pyparsing
ruamel.yaml
ruamel.yaml.clib
uritemplate

cloudinary

affinda
pandas
thefuzz
python-Levenshtein
